Factor out endpoints into the beginnings of a general IntList class
